- 변수명 정리 - 메소드 분리 진행하였습니다~!
