+++ src/main/java/Calculator.java
@@ -1,31 +1,13 @@
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
 
 public class Calculator {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        String input = sc.nextLine();
-        if (input.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
 
-        String[] ar = input.split(" ");
-        List<Integer> nums = new ArrayList<>();
-        List<String> calculate = new ArrayList<>();
-
-        for (int i = 0; i < ar.length; i++) {
-            if (i % 2 == 0) {
-                nums.add(Integer.parseInt(ar[i]));
-            } else {
-                calculate.add(ar[i]);
-            }
-        }
-
+    // 계산
+    public int calculate(List<Integer> nums, List<String> operators) {
         int result = nums.get(0);
-        for(int i = 0; i < calculate.size(); i++) {
-            switch (calculate.get(i)) {
+        for(int i = 0; i < operators.size(); i++) {
+            switch (operators.get(i)) {
                 case "+":
                     result = addition(result, nums.get(i+1));
                     break;
@@ -42,7 +24,7 @@ public class Calculator {
                     throw new IllegalArgumentException();
             }
         }
-        System.out.println(result);
+        return result;
     }
 
     public static int addition(int num, int result) {
